+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## Unreleased
+
+### Added
+
+- Go to definition and Peek definition can resolve TEMPLATE actions to their C# source if Daggerfall Unity project is part of current workspace.
+- Added action `Find action references` that finds TEMPLATE references from a C# action definition if Daggerfall Unity project is part of current workspace.
+
 ## 0.12.0
 
 ### Added